+++ internal/repository/postgresql/migrations/000008_alter_feed_subscriptions_add_alias_column.up.sql
@@ -1,0 +1,5 @@
+-- Copyright (c) VirtualTam
+-- SPDX-License-Identifier: MIT
+
+ALTER TABLE feed_subscriptions
+ADD COLUMN alias TEXT NOT NULL DEFAULT '';